Restructured
Deleted csvscrims and moved its content into Main. The plain text email system is working somewhat at this point.
